Initial info tree and material=>recycling info conversion object
nodeList and whereDoesItGo files only contain Array of objects used in infoTreeLogic page. That page converts one of these objects (currently the one housed in whereDoesItGo) and converts it into a tree structure, with child=>parent and parent=> child relationships. I've left comments in the code describing the specific functions and progress that can be made in the future