Refactored createRecycleTree and Added data flow structure
Renamed recycleBuddyScratch to createRecycleTree. Refactored createRecycleTree to use forEach loops and clean up a lot of other aspects of the code. Increased quality of comments. SeattleGovScrape serves as a prototype space for scrapeHtml function. createDataStore demonstrates how data will flow. This file is mostly a skeleton, with many TODOs and empty functions