Initial info tree and material=>recycling info conversion object (#10)
* Initial info tree and material=>recycling info conversion object

nodeList and whereDoesItGo files only contain Array of objects used in infoTreeLogic page. That page converts one of these objects (currently the one housed in whereDoesItGo) and converts it into a tree structure, with child=>parent and parent=> child relationships. I've left comments in the code describing the specific functions and progress that can be made in the future

* Refactored createRecycleTree and Added data flow structure

Renamed recycleBuddyScratch to createRecycleTree. Refactored createRecycleTree to use forEach loops and clean up a lot of other aspects of the code. Increased quality of comments. SeattleGovScrape serves as a prototype space for scrapeHtml function. createDataStore demonstrates how data will flow. This file is mostly a skeleton, with many TODOs and empty functions

* Initial info tree and material=>recycling info conversion object

nodeList and whereDoesItGo files only contain Array of objects used in infoTreeLogic page. That page converts one of these objects (currently the one housed in whereDoesItGo) and converts it into a tree structure, with child=>parent and parent=> child relationships. I've left comments in the code describing the specific functions and progress that can be made in the future